Remove Developer Blog Tutorial under JavaScript -> Remix (#11953)
* Remove Developer Blog Tutorial under JavaScript -> Remix

This pull request removes a resource link that is no longer accessible (404 error). Keeping the list up-to-date and free of broken links ensures a better experience for readers and maintains the overall quality of the repository.

* Remove `Jokes-App-Tutorial` under JavaScript -> Remix

This pull request removes a [Jokes App Tutorial](https://remix.run/docs/en/v1/tutorials/jokes), which is no longer accessible.

* Remove Remix from the index